feature: "Flatten" the ExprKind::Number variants, replacing it with ExprKind::{Integer, Real}
These are structurally different types of expressions. They *can* be semantically
grouped as numbers, but ExprKind is a structural representation, and should not
group them that way.

This is also part of the argument against having an ExprKind::Atom variant
that contains Symbol, String, etc. variants.

This is also easier to match over without extra nesting.

If you want to check if an Expr is a number, use the new Expr::try_number() method.