feature: Add unstable `ExprRefCmp` type (#12)
* Revert "cleanup: Remove ExprRefCmp"

This reverts commit 7ce91679f5e877320f5b165d2525f63d88f61697.

* docs: Fix cargo doc warning about non-link hyperlinks in README.md

* feature: Re-add ExprRefCmp type behind the "unstable_parse" feature flag

This type is useful for the source mapping feature of wl-parse, but
it is not clear that this type is more generally useful, or that its
design is final.

Wrt. possible alternative designs, in the source mapping use case, the
general goal is to associate a piece of data (the source map info) with a
specific Expr instance. The ExprRefCmp type makes it easier to store that
data external to the Expr instance itself, in a global hashmap.

But an alternative possible design might be to support a way to embed the
data in the Expr itself, using e.g. a formal type parameter, or a `typemap`
like datastructure in the Expr instance itself.